Rename GRPO Trainer
