+++ src/open_r1/trainer/__init__.py
@@ -1,4 +1,4 @@
-from .gpro_trainer import Qwen2VLGRPOTrainer
+from .grpo_trainer import Qwen2VLGRPOTrainer
 
 
 __all__ = ["Qwen2VLGRPOTrainer"]
